+++ HISTORY.md
@@ -9,6 +9,7 @@
     * typing strings into the search-field again shows relevant blocks (regression from IME)
     * fixed project dialog's search-field behevior (regression from IME)
     * morphic collision detection off-by-1 fix, thanks, Dariusz!
+    * fixed MY PARTS so mutating the result list has no effect
     * fixed a typo in the OF-reporter's help screen, thanks, @jasonappah
     * enable costumes created in the vector editor to be stretchable in Firefox, thanks, @coproc
 * **Translation Updates:**
@@ -21,6 +22,7 @@
 * morphic: new "reactToInput" text-editing event
 * objects: fixed #2485 (find blocks and text-entry mode feature)
 * gui: fixed ProjectDialog's search field behavior for IME 
+* threads: fixed MY PARTS so mutating the result list has no effect
 
 ### 2019-10-11
 * objects, threads: accept a list of pixels in the SWITCH TO COSTUME block